+++ src/lib.rs
@@ -155,6 +155,7 @@ mod serde_util;
 pub use key::SecretKey;
 pub use key::PublicKey;
 pub use key::ONE_KEY;
+pub use key::KeyPair;
 pub use context::*;
 use core::marker::PhantomData;
 use core::{mem, fmt, ptr, str};
